+++ compiler/src/iree/compiler/Codegen/Dialect/IREECodegenDialect.cpp
@@ -7,6 +7,7 @@
 #include "iree/compiler/Codegen/Dialect/IREECodegenDialect.h"
 
 #include "iree/compiler/Codegen/Dialect/IREECodegenDialect.cpp.inc"
+#include "iree/compiler/Codegen/Dialect/IREECodegenOps.h"
 #include "iree/compiler/Codegen/Dialect/LoweringConfig.h"
 #include "iree/compiler/Codegen/Dialect/UKernelOps.h"
 #include "mlir/IR/DialectImplementation.h"
@@ -37,6 +38,10 @@ void IREECodegenDialect::initialize() {
   initializeCodegenAttrs();
   addInterfaces<IREECodegenDialectOpAsmInterface>();
 
+  addOperations<
+#define GET_OP_LIST
+#include "iree/compiler/Codegen/Dialect/IREECodegenOps.cpp.inc"
+      >();
   addOperations<
 #define GET_OP_LIST
 #include "iree/compiler/Codegen/Dialect/UKernelOps.cpp.inc"
