+++ compiler/src/iree/compiler/Codegen/LLVMCPU/LLVMCPUMaterializeEncodingPass.cpp
@@ -7,6 +7,7 @@
 #include "iree-dialects/Dialect/LinalgExt/IR/LinalgExtDialect.h"
 #include "iree-dialects/Dialect/LinalgExt/Passes/Passes.h"
 #include "iree/compiler/Codegen/Common/EncodingInfo.h"
+#include "iree/compiler/Codegen/Dialect/IREECodegenDialect.h"
 #include "iree/compiler/Codegen/LLVMCPU/Utils.h"
 #include "iree/compiler/Codegen/PassDetail.h"
 #include "iree/compiler/Dialect/Flow/IR/FlowOps.h"
@@ -100,9 +101,10 @@ static MatmulTileParams chooseMatmulTileParams(MatmulType type,
 struct LLVMCPUMaterializeEncodingPass
     : public LLVMCPUMaterializeEncodingBase<LLVMCPUMaterializeEncodingPass> {
   void getDependentDialects(DialectRegistry &registry) const override {
-    registry.insert<arith::ArithDialect, affine::AffineDialect,
-                    IREE::Flow::FlowDialect,
-                    IREE::LinalgExt::IREELinalgExtDialect>();
+    registry
+        .insert<arith::ArithDialect, affine::AffineDialect,
+                IREE::Flow::FlowDialect, IREE::LinalgExt::IREELinalgExtDialect,
+                IREE::Codegen::IREECodegenDialect>();
   }
   void runOnOperation() override;
 };
