+++ compiler/src/iree/compiler/Codegen/Transforms/BUILD.bazel
@@ -24,6 +24,7 @@ iree_compiler_cc_library(
     ],
     deps = [
         "//compiler/src/iree/compiler/Codegen/Utils",
+        "//compiler/src/iree/compiler/Codegen/Dialect:IREECodegenDialect",
         "//compiler/src/iree/compiler/Dialect/Flow/IR",
         "//compiler/src/iree/compiler/Dialect/HAL/IR",
         # TODO(#13181) : Remove this dependency on VMVX dialect.
