+++ compiler/src/iree/compiler/Codegen/Transforms/CMakeLists.txt
@@ -35,6 +35,7 @@ iree_cc_library(
     MLIRTransforms
     MLIRValueBoundsOpInterface
     MLIRVectorDialect
+    iree::compiler::Codegen::Dialect::IREECodegenDialect
     iree::compiler::Codegen::Utils
     iree::compiler::Dialect::Flow::IR
     iree::compiler::Dialect::HAL::IR
