+++ compiler/src/iree/compiler/Dialect/VMVX/Conversion/VMVXToVM/test/BUILD.bazel
@@ -19,7 +19,6 @@ iree_lit_test_suite(
             "binary.mlir",
             "copy.mlir",
             "fill.mlir",
-            "query_tile_sizes.mlir",
             "unary.mlir",
         ],
         include = ["*.mlir"],
