+++ compiler/src/iree/compiler/Dialect/VMVX/Conversion/VMVXToVM/test/CMakeLists.txt
@@ -17,7 +17,6 @@ iree_lit_test_suite(
     "binary.mlir"
     "copy.mlir"
     "fill.mlir"
-    "query_tile_sizes.mlir"
     "unary.mlir"
   TOOLS
     FileCheck
