+++ compiler/src/iree/compiler/Dialect/VMVX/IR/VMVXOps.td
@@ -75,27 +75,6 @@ def VMVX_GetRawInterfaceBindingBufferOp : VMVX_PureOp<
   }];
 }
 
-def VMVX_QueryTileSizesOp : VMVX_PureOp<"query_tile_sizes", [Pure]> {
-  let summary = "Returns tile sizes to use to materialize the given encoding, based on runtime CPU capabilities.";
-  let description = [{
-    TODO write me
-  }];
-
-  let arguments = (ins
-    Variadic<VMVX_Index>:$sizes,
-    I32Attr:$flags
-  );
-  let results = (outs
-    Variadic<VMVX_Index>:$tileSizes
-  );
-
-  let assemblyFormat = [{
-    `sizes` `` `(` $sizes `)`
-    `flags` `` `(` $flags `)`
-    `->` type(results) attr-dict
-  }];
-}
-
 //===----------------------------------------------------------------------===//
 // VMVX Ops: ABI
 //===----------------------------------------------------------------------===//
